src/requirements.txt
PyQt6==6.5.0
pyserial==3.5
PySide6==6.5.0
